# api/openapi/portal-bff.yaml
openapi: 3.0.3
info:
  title: token-gateway Portal Web BFF
  version: 0.2.0
  description: Browser-facing Portal BFF contract owned by cmd/console.
servers:
  - url: http://localhost:9505
    description: Local console process
...
paths:
...
  /api/portal/v1/tasks:
    get:
      tags:
        - Portal Web Tasks
      summary: List project-scoped tasks
      operationId: listPortalTasks
      security:
        - portalSession: []
      parameters:
        - name: api_key_id
          in: query
          required: false
          schema:
            type: string
        - name: request_id
          in: query
          required: false
          schema:
            type: string
        - name: model
          in: query
          required: false
          schema:
            type: string
        - name: provider_type
          in: query
          required: false
          schema:
            type: string
        - name: channel_id
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: cursor
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: Project-scoped tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
  /api/portal/v1/tasks/{task_id}:
    get:
      tags:
        - Portal Web Tasks
      summary: Get one project-scoped task
      operationId: getPortalTask
      security:
        - portalSession: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project-scoped task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskObject"
        "401":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /api/portal/v1/tasks/{task_id}/cancel:
    post:
      tags:
        - Portal Web Tasks
      summary: Reserved task cancel endpoint
      operationId: cancelPortalTask
      security:
        - portalSession: []
          csrfHeader: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "501":
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    portalSession:
      type: apiKey
      in: cookie
      name: tg_portal_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  responses:
    ErrorResponse:
      description: Stable console error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    TaskListResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TaskObject"
        next_cursor:
          type: string
          nullable: true
    TaskObject:
      type: object
      required:
        - id
        - object
        - status
        - type
        - created
        - progress
      properties:
        id:
          type: string
        object:
          type: string
          enum:
            - image.generation.task
            - image.edit.task
            - video.generation.task
            - audio.generation.task
            - music.generation.task
        created:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        request_id:
          type: string
        api_key_id:
          type: string
        model:
          type: string
        status:
          type: string
          enum:
            - pending
            - queued
            - processing
            - completed
            - failed
            - cancelled
            - expired
        progress:
          type: integer
          minimum: 0
          maximum: 100
        type:
          type: string
        provider_task_id:
          type: string
        provider_type:
          type: string
        channel_id:
          type: string
        results:
          type: array
          items:
            type: string
        assets:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
              type:
                type: string
              mime_type:
                type: string
              provider:
                type: string
              expires_at:
                type: string
                format: date-time
              metadata:
                type: object
                additionalProperties:
                  type: string
        provider_metadata:
          type: object
          additionalProperties:
            type: string
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            type:
              type: string
            retryable:
              type: boolean
        task_info:
          type: object
          properties:
            can_cancel:
              type: boolean
            estimated_time:
              type: integer
            video_duration:
              type: integer
        usage:
          type: object
          properties:
            billing_rule:
              type: string
            input_tokens:
              type: integer
            output_tokens:
              type: integer
            total_tokens:
              type: integer
        metadata:
          type: object
          additionalProperties:
            type: string
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      required:
        - code
        - message
        - type
      properties:
        code:
          type: string
          example: unauthorized
        message:
          type: string
        type:
          type: string
        retryable:
          type: boolean
        request_id:
          type: string